refactor(release): split Fact projection verification
